Move IReactProps to Utilities package.
