+++ change/@fluentui-react-4937b01e-1cf3-4d3a-bd5e-adccb2ba8c96.json
@@ -1,0 +1,7 @@
+{
+  "type": "none",
+  "comment": "chore: Removing deprecated React.Props from DelayedRender.",
+  "packageName": "@fluentui/react",
+  "email": "leon.chi@displayr.com",
+  "dependentChangeType": "none"
+}
